sensor/requirements.txt
kafka-python-ng==2.2.2
psutil==5.9.8
